splited packager.py
